New Methods and Bug fixes
All of the methods were untested and were to say, a rough draft. I went over and tested them, fixed what needed to, and did the same for the new methods I had written up, such as list(), edit(), copyTo(), and copyAs(). Unless I skimmed over something, this should be all for this class.