Fixing errors and improving efficiency
I was under the assumption .isFile() from Java's Utils for some reason determined if its a file or not depending on whether it has a file extension indicated by the period, but if this file doesn't exists it returns true. This means whenever i tried to create a file, it would instead create a folder because my create method used isFile to determine whether creating a folder or file. I simply split the method to createFolder() and create()

Also, some of the methods using array parameters have been made into Lists that way during its check it can remove the element found and decrease the amount of elements to check for during each line.